Tidied up and linting
